Fix some clang-tidy bugprone-narrowing-conversions warnings

+++ src/sound/SoundDevice.cc
@@ -8,6 +8,7 @@
 #include "MemBuffer.hh"
 #include "MSXException.hh"
 #include "aligned.hh"
+#include "narrow.hh"
 #include "ranges.hh"
 #include "one_of.hh"
 #include "vla.hh"
@@ -44,8 +45,8 @@ static void allocateMixBuffer(size_t size)
 void SoundDevice::addFill(float*& buf, float val, unsigned num)
 {
 	// Note: in the past we tried to optimize this by always producing
-	// a multiple of 4 output values. In the general case a sounddevice is
-	// allowed to do this, but only at the end of the soundbuffer. This
+	// a multiple of 4 output values. In the general case a SoundDevice is
+	// allowed to do this, but only at the end of the sound buffer. This
 	// method can also be called in the middle of a buffer (so multiple
 	// times per buffer), in such case it does go wrong.
 	assert(num > 0);
@@ -89,7 +90,7 @@ float SoundDevice::getAmplificationFactorImpl() const
 void SoundDevice::registerSound(const DeviceConfig& config)
 {
 	const auto& soundConfig = config.getChild("sound");
-	float volume = soundConfig.getChildDataAsInt("volume", 0) / 32767.0f;
+	float volume = narrow<float>(soundConfig.getChildDataAsInt("volume", 0)) / 32767.0f;
 	int devBalance = 0;
 	std::string_view mode = soundConfig.getChildData("mode", "mono");
 	if (mode == "mono") {
@@ -318,7 +319,7 @@ bool SoundDevice::mixChannels(float* dataOut, unsigned samples)
 
 	// In the past we had ARM and x86-SSE2 optimized assembly routines for
 	// the stuff below. Currently this code is only rarely used anymore
-	// (only when recording or muting individual soundchip channels), so
+	// (only when recording or muting individual sound chip channels), so
 	// it's not worth the extra complexity anymore.
 	unsigned num = samples * stereo;
 	unsigned i = 0;
